add file populer_content_rec.py
The popular_content_rec.py file is a content recommendation algorithm for the most popular content which is sorted based on the popularity of the content from clicks and likes. This file has input in the form of rating_dir and fiction_dir with output of the 20 most popular fanfictions.